[deckhouse] Move the operations package under the operation controller
